requirements.txt
adium-theme-ubuntu==0.3.4
django-markdown-deux==1.0.5
markdown2==2.3.4
pytz==2017.2
unity-lens-photos==1.0
virtualenv==15.1.0
